% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-group-compositions.R
\name{build_contrast_specifications}
\alias{build_contrast_specifications}
\title{Build contrast specifications from group levels and custom overrides.}
\usage{
build_contrast_specifications(group, custom_contrasts = NULL)
}
\description{
Build contrast specifications from group levels and custom overrides.
}
\keyword{internal}
