% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hellinger-internals.R
\name{build_distance_matrix_from_draw}
\alias{build_distance_matrix_from_draw}
\title{Build distance matrix from one posterior draw of compositions.}
\usage{
build_distance_matrix_from_draw(draw_matrix)
}
\arguments{
\item{draw_matrix}{Subjects-by-categories frequency matrix (one draw).}
}
\value{
Symmetric Hellinger distance matrix.
}
\description{
Build distance matrix from one posterior draw of compositions.
}
\keyword{internal}
