% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hellinger-internals.R
\name{compute_hellinger_distance_freqs}
\alias{compute_hellinger_distance_freqs}
\title{Hellinger distance between two frequency vectors on a simplex.}
\usage{
compute_hellinger_distance_freqs(freqs1, freqs2)
}
\arguments{
\item{freqs1}{Numeric frequency vector (sums to 1).}

\item{freqs2}{Numeric frequency vector (sums to 1).}
}
\value{
Scalar Hellinger distance in \link{0, 1}.
}
\description{
Hellinger distance between two frequency vectors on a simplex.
}
\keyword{internal}
