% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-group-compositions.R
\name{run_contrast_permutation}
\alias{run_contrast_permutation}
\title{Run permutation null for one contrast.}
\usage{
run_contrast_permutation(
  counts,
  group,
  spec,
  n_permutations,
  prior_pseudocounts
)
}
\description{
Run permutation null for one contrast.
}
\keyword{internal}
