% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-group-compositions.R
\name{run_permutation_enrichment}
\alias{run_permutation_enrichment}
\title{Execute permutation enrichment across all contrasts.}
\usage{
run_permutation_enrichment(
  composition,
  contrast_specs,
  n_permutations,
  n_cores,
  prior_pseudocounts
)
}
\description{
Execute permutation enrichment across all contrasts.
}
\keyword{internal}
