% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-group-compositions.R
\name{run_permutation_worker}
\alias{run_permutation_worker}
\title{Parallel worker wrapper for permutation contrasts.}
\usage{
run_permutation_worker(spec, counts, group, n_permutations, prior_pseudocounts)
}
\description{
Parallel worker wrapper for permutation contrasts.
}
\keyword{internal}
