% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-group-compositions.R
\name{subset_for_contrast}
\alias{subset_for_contrast}
\title{Subset count matrix and distance inputs for a contrast.}
\usage{
subset_for_contrast(counts, group, spec)
}
\description{
Subset count matrix and distance inputs for a contrast.
}
\keyword{internal}
