% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-group-compositions.R
\name{validate_composition_for_inference}
\alias{validate_composition_for_inference}
\title{Validate that groups have enough subjects for pairwise distances.}
\usage{
validate_composition_for_inference(composition)
}
\description{
Validate that groups have enough subjects for pairwise distances.
}
\keyword{internal}
