% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-group-compositions.R
\name{permutation_p_value}
\alias{permutation_p_value}
\title{Permutation p-value for one contrast (includes observed in numerator).}
\usage{
permutation_p_value(observed_ratio, permuted_ratios)
}
\description{
Permutation p-value for one contrast (includes observed in numerator).
}
\keyword{internal}
